chatbot/requirements.txt
flask==3.0.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
requests==2.32.3
pdfplumber==0.11.0
rank-bm25==0.2.2
scikit-learn==1.5.1
python-dotenv==1.0.1
redis==5.0.8
